Typecast compar warnings in contextmenu.c

+++ src/ui/contextmenu.c
@@ -29,7 +29,7 @@ static void calculate_pos_and_width(CONTEXTMENU *b, int *x, int *w) {
     }
 
     // Push away from the right border to fit.
-    if (*x + *w >= settings.window_width) {
+    if (*x + *w >= (int)settings.window_width) {
         *x -= *w;
     }
 }
@@ -149,7 +149,7 @@ void contextmenu_new_ex(uint8_t count, void *userdata, void (*onselect)(uint8_t)
 
     b->y      = mouse.y;
     b->height = CONTEXT_HEIGHT * count;
-    if (b->y + b->height >= settings.window_height) {
+    if (b->y + b->height >= (int)settings.window_height) {
         b->y -= b->height;
     }
     b->x     = mouse.x;
